backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
billiard==4.2.4
bleach==6.3.0
build==1.3.0
cachetools==6.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
coverage==7.13.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
dparse==0.6.4
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
Faker==20.1.0
fastapi==0.127.1
fastapi-cache2==0.2.1
filelock==3.20.1
flatbuffers==25.12.19
flower==2.0.1
freezegun==1.4.0
frozenlist==1.8.0
fsspec==2025.12.0
google-auth==2.45.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
humanize==4.15.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
kubernetes==34.1.0
lyricsgenius==3.7.5
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nltk==3.9.2
numpy==2.4.0
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions
orjson==3.11.5
overrides==7.7.0
packaging==25.0
passlib==1.7.4
pdfminer.six
pdfplumber==0.11.8
pendulum==3.1.0
Pillow==10.1.0
pluggy==1.6.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.5.0
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
python-dateutil==2.9.0
python-dotenv==1.2.1
python-engineio==4.13.0
python-jose==3.5.0
python-multipart==0.0.21
python-socketio==5.16.0
pytz==2025.2
PyYAML==6.0.3
redis==4.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
responses==0.24.1
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruff==0.14.13
safety==3.7.0
safety-schemas==0.0.16
sentry-sdk==2.49.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
spotipy==2.25.2
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tomlkit==0.14.0
tornado==6.5.4
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
